Created GET /employees endpoint
