+++ api/middlewares/roleBasedAccessMiddleware.js
@@ -1,10 +1,11 @@
 const messages = require('../utils/messages');
 const generateResponseBody = require('../utils/responseGenerator');
 const winston = require('../utils/winston');
+const constants = require('../utils/constants');
 
 const allowAccess = (requiredRole) => {
     return (req, res, next) => {
-        if (requiredRole.includes(req.authorizedUser.role)) {
+        if (req.authorizedUser.role === constants.userRoles.Admin || requiredRole.includes(req.authorizedUser.role)) {
             next();
         } else {
             winston.info(`Unauthorized access blocked`, { req });
