+++ api/routes/controllers/userController.js
@@ -70,16 +70,6 @@ const DeleteUser = (req, res) => {
   return userHelper.DeleteUser(req, res);
 }
 
-/**
- * Function to create employee
- * @param {*} req
- * @param {*} res
- * @returns {}
- */
-const CreateEmployee = (req, res) => {
-  return userHelper.CreateEmployee(req, res);
-}
-
 module.exports = {
   GetAllUsers,
   GetUsersPendingApproval,
@@ -88,5 +78,4 @@ module.exports = {
   SuspendUser,
   GetDeletedUsers,
   DeleteUser,
-  CreateEmployee,
 };
