+++ api/routes/index.js
@@ -1,5 +1,6 @@
 var express = require('express');
 var usersRouter = require('./usersRoute');
+var employeesRouter = require('./employeesRoute');
 var authRouter = require('./authRoute');
 var propertiesRouter = require('./propertiesRoute');
 var complaintsRouter = require('./complaintsRoute');
@@ -10,6 +11,7 @@ var router = express.Router();
 router.use('/auth', authRouter);
 router.use(authenticateToken);
 router.use('/user', usersRouter);
+router.use('/employees', employeesRouter)
 router.use('/properties', propertiesRouter);
 router.use('/complaints', complaintsRouter);
